3_final_project/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
sqlalchemy
# psycopg2-binary
python-dotenv
nanoid
# re

pydantic[email]
